+++ lib/Echidna.hs
@@ -23,8 +23,10 @@ import Echidna.Types.Tx
 import Echidna.Types.World
 import Echidna.Transaction
 import Echidna.Processor
+import Echidna.Output.Source
 
 import qualified Data.List.NonEmpty as NE
+import qualified Data.Text.IO as T (putStrLn)
 
 -- | This function is used to prepare, process, compile and initialize smart contracts for testing.
 -- It takes:
